Improving Modularity of Program
