+++ render_css/default.css
@@ -1,5 +1,5 @@
 .container {
     display:grid;
     grid-template-rows: 1fr 1fr 1fr;
-    grid-template-columns: 15% 1fr 15%;
+    grid-template-columns: 10% 1fr 10%;
 }
