+++ render_css/mainmenu.css
@@ -1,0 +1,19 @@
+#title {
+    grid-column:2;
+}
+.bottom-button{
+    display:flex;
+    flex-direction:row;
+    grid-column:2;
+    grid-row:3;
+}
+
+.bottom-button button{
+    margin:5px;
+}
+
+#analyse-form {
+    margin: 10px 10px 10px 10px;
+    grid-column:2;
+    grid-row:2;
+}
