--- render_css/result-table.css
@@ -1,2 +1,0 @@
-.container {
-}
