+++ renderer/credentials.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html>
 <head>
     <meta charset="UTF-8">
     <!-- https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP -->
@@ -7,17 +7,15 @@
     <meta http-equiv="X-Content-Security-Policy" content="default-src 'self'; script-src 'self'">
     <link rel="stylesheet" type="text/css" href="../node_modules/bootstrap/dist/css/bootstrap-grid.min.css">
     <link rel="stylesheet" type="text/css" href="../node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <title>Transcription Service</title>
-    <link rel="stylesheet" type="text/css" href="../render_css/result-table.css">
+    <link rel="stylesheet" type="text/css" href="../render_css/default.css">
+    <title>Service Credentials</title>
 </head>
 <body>
 <div class="container">
-    <div id="tableresult">
-
-    </div>
+    <h1 id="title">Service Credentials</h1>
 </div>
 <!-- You can also require other files to run in this process -->
 <script src="../renderer.js"></script>
-<script src="result-table.js"></script>
+<script src="loginform.js"></script>
 </body>
 </html>
